Variables for storing student data
nameLine= This variable is intended to store the name of a student. numClassesString= This variable is intended to store the number of classes a student is enrolled in. It's declared as a String, which suggests that the number of classes might initially be read or as text from a file or user input.  studentNumber= This is  used to store a student's identification number.