Valid data from the input fiel
Passing four pieces of student data: firstName, secondName, studentNumber, and numClassesString. If the data is valid, it is expected to be "Valid" and output to the appropriate format to the output file. Also String workload = getWorkloadDescription(Integer.parseInt(numClassesString)); = calls another method getWorkloadDescription to pass/converted from a string to an integer based on number of classes this will be set up in the later stage.  writer.writer, writes student data that is formatted to the output file, writer.newLine will make it more readable if there is more then one students in the output file it will create extra break in between students.  If validationResult does not equal "Valid" it will print out error message to the console.