Relevant pieces of information are going to be tested
In this part of the code we are going to test firstName, secoundName, studentNumber and numClassesString. Based on the return we get Valid or Invalid program will decide if to continue with the data or to display relevant error message that we have set up. if (!Pattern.matches("[a-zA-Z]+", firstName)) this is same for secoundName as-well. Pattern matches method takes regex and input and returns true if the input matches the pattern. [a-zA-Z]+ is in place to specify what pattern the frist and secound name should match: a-z= lowercase alphabet, A-Z= uppercase alphabet. The += First/secound name must have one or more letters. If true which means that the input doesnt match that pattern error message will be displayed. \\d{2}[a-zA-Z]{2,3}[0-9]+ this will help us with getting the correct student number format. d{2}=two digits must be present. [a-zA-Z]{2,3}=This matches between 2 and 3 alphabetical characters, either lowercase or uppercase. {2,3} quantifier indicates the minimum and maximum number of characters that must be present from this range. [0-9]+= This matches one or more digits. [0-9] specifies any digit, and the +  means one or more. Again if invalid input error message will appear. import java.util.regex.Pattern= had to be imported for us to be able to use Java Regular Expression (Regex) which allows us to do pattern matching.