Turning string into an integer
numClassesString indeed represents integer that falls within an acceptable range for the number of classes. if numClasses makes sure that no. of classes falls in between 1 and 8. catch (NumberFormatException e) runs when Integer.parseInt(numClassesString) fails. Exception occurs when numClassesString is not in correct format therefore the error message is displayed.