Number of classes validation
getWorkloadDescription, is designed to categorise the workload of a student based on the number of classes they are enrolled in. We are taking numClasses representing the total number of classes. Based on this we will return a String that describes the student's workload. If (numClasses == 1) their workload is considered "Very Light". Else if (numClasses == 2) their workload is described as "Light". else if (numClasses >= 3 && numClasses <= 5) students taking between 3 and 5 classes the workload is categorised as "Part Time". Use of && make sure this condition is true only if numClasses falls within this range. else any other number 6 or more is categorized as "Full Time"