Ensure that the student number year is at least 2020 (i.e. that the number starts with 20 or higher)
To achieve student number that starts with the year 2020 or later, I have changed the regex expression aka updated it so it works for the specific criteria. 20\\d{1}= This part focuses on the "20" and its followed by exactly one digit \\d{1} . Another part that has to be changed of this expression is \\d+ which means that that the remaining portion of the student number should be made up of one more digits.