Amended Ensure that the student number year is at least 2020
In my previous attend to make this work I have wrongly asked the user to include 20 and any other digit which makes it three digits wheras it is meant to be 20 or higher. if (!Pattern.matches("\\d{2}[a-zA-Z]{2,3}\\d+", studentNumber)) {= asking the input file to have 2 any digits loosly. Integer.parseInt(studentNumber.substring(0, 2)) checks the first part which is two characters of the student number which represents the year and then if these 2 figures are less than 20 then the error message is displayed.